WWT/pyproject.toml
[project]
name = "question-driven-analytics"
version = "0.1.0"
description = "Question-Driven Analytics Platform POC - Ask natural language questions about supply chain data"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "google-generativeai>=0.8.0",
    "duckdb>=1.0.0",
    "httpx>=0.27.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.6.0",
    "python-multipart>=0.0.9",
]

[project.scripts]
qda = "src.main:run"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
